requirements.txt
Django~=4.1.0
djangorestframework~=3.13.1
selenium==4.4.0
bs4~=0.0.1
requests~=2.28.1
beautifulsoup4~=4.11.1
jellyfish~=0.9.0